Ensure 50% & 50 % are treated as two segments in TTS
The norm variant is kept out of TTS with a filter tag, the err-orth variant (without the space) is split in two: one for normal analysers, and one for TTS which forces a split-up in two segments.

For this to work properly, we need to filter out the TTS tags/strings from all other fst’s.