TTS work on giellalt/lang-sme#499  - not solved
It is progress in the sense that all digits and symbols are now converted to text, but it is a step backwards in the sense that date ranges are not analysed as dates at all.

But reading each digit and symbol aloud separately is definitely better than silence, as it used to be.